refactor(channex): extract certification diagnostics services
